refactor: TodoList 컴포넌트 VAC 패턴 적용 #10
