+++ client/src/components/todos/TodoList/index.tsx
@@ -2,34 +2,26 @@ import { startTransition } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import { useGetTodosQuery } from '../../../queries/todo';
-import Button from '../../common/Button';
-import { Container, Wrapper, Text, Todos, Todo, Title, Content } from './style';
-import todoType from '../../../types/todos';
+import TodoData from '../../../types/todos';
+import TodoListView from './TodoListView';
+
+export interface TodoListViewProps {
+  todos: TodoData[];
+  navigateToNew: () => void;
+  navigateToDetail: (id: string) => void;
+}
 
 const TodoList = () => {
   const navigate = useNavigate();
-
   const { data } = useGetTodosQuery();
 
-  return (
-    <Container>
-      <h3 className='sr-only'>할 일 목록</h3>
-      <Wrapper>
-        <Text>Tasks {data?.data.data.length || 0}</Text>
-        <Button type='button' size='small' onClick={() => navigate('/new')}>
-          NEW
-        </Button>
-      </Wrapper>
-      <Todos>
-        {data?.data.data.map((todo: todoType) => (
-          <Todo key={todo.id} onClick={() => startTransition(() => navigate(`/todos/${todo.id}`))}>
-            <Title>{todo.title}</Title>
-            <Content>{todo.content}</Content>
-          </Todo>
-        ))}
-      </Todos>
-    </Container>
-  );
+  const todoListViewProps: TodoListViewProps = {
+    todos: data?.data.data,
+    navigateToNew: () => navigate('/new'),
+    navigateToDetail: (id) => startTransition(() => navigate(`/todos/${id}`)),
+  };
+
+  return <TodoListView {...todoListViewProps} />;
 };
 
 export default TodoList;
